Added IR sensors for motor base and differnt modes
Added the IR sensors to the base of the arm so that we can tell
what position the arm is in. I also added code for the operator to
switch modes. One mode will be the standard operator mode which has
been already implemented and a auto mode where we use the IR sensors.